numapar: context cancellation/failure propagation and serial lifecycle
The terminal-status machinery, its cross-context propagation and the
machine-parseable serial lifecycle events, all inactive until the
NUMA_PAR switch-on:

- context_record_failure()/load_test_status() operate on the atomic
  per-context status (first terminal reason wins); the UI-requested
  bail is published as a cross-context cancellation request and every
  team converges on the same exit path at its next safe point
  (test_publish_cancel_request()/test_request_cancel()/
  test_stop_requested(), BAILOUT uses the latter in all tests).
- WAVE_FINISH promotes fatal context statuses to the stage-global
  failure latch before the next wave rebinds the reusable contexts, so
  an early-wave failure is never erased, and reports every terminal
  status with its wave/domain identity; a stage failure disables
  NUMA_PAR and restarts the run in legacy mode rather than silently
  omitting memory.
- serial_log_wave_bind()/serial_log_context_end() keep the serial log
  unambiguous with concurrent contexts.
- Global master rotation is gated out of NUMA_PAR runs; each context
  keeps its fixed master.

+++ system/smp.c
@@ -1344,7 +1344,7 @@ void map_the_numa_memory_range(uint8_t highest_bit)
     //
     // First step, map the pm_map.
     //
-    for (i = 0; i < (unsigned int)pm_map_size; i++) {
+    for (i = 0; i < pm_map_size; i++) {
         node_nu = (pm_map[i].start >> (node_offset - PAGE_SHIFT)) & 0xF;
         if (node_nu != 0) {
             pm_map[i].start &= ~((uint64_t)0xF << (node_offset - PAGE_SHIFT));
@@ -1358,7 +1358,7 @@ void map_the_numa_memory_range(uint8_t highest_bit)
     //
     // Second step, map the memory_affinity_ranges.
     //
-    for (i = 0; i < (unsigned int)num_memory_affinity_ranges; i++) {
+    for (i = 0; i < num_memory_affinity_ranges; i++) {
         if (memory_affinity_ranges[i].proximity_domain_idx != 0) {
             node_nu = (memory_affinity_ranges[i].start >> node_offset) & 0xF;
             if (node_nu != 0) {
@@ -1382,7 +1382,7 @@ void check_if_needs_to_map(void)
     if (num_proximity_domains == 0x0) {
         return;
     } else {
-        for (i = 0; i < (unsigned int)num_memory_affinity_ranges; i++) {
+        for (i = 0; i < num_memory_affinity_ranges; i++) {
             if (memory_affinity_ranges[i].proximity_domain_idx != 0) {
                 local_memory_area_bits = checkout_max_memory_bits_of_this_numa_node(i);
                 if (max_memory_bits < local_memory_area_bits) {
